+++ wandb/sdk/wandb_run.py
@@ -4,9 +4,7 @@ from enum import IntEnum
 import glob
 import json
 import logging
-import numbers
 import os
-import platform
 import re
 import sys
 import threading
@@ -28,9 +26,8 @@ from typing import (
 )
 from typing import TYPE_CHECKING
 
-import click
 import requests
-from six import iteritems, string_types
+from six import string_types
 from six.moves import _thread as thread
 from six.moves.collections_abc import Mapping
 from six.moves.urllib.parse import quote as url_quote
@@ -42,14 +39,14 @@ from wandb._globals import _datatypes_set_callback
 from wandb.apis import internal, public
 from wandb.apis.public import Api as PublicApi
 from wandb.proto.wandb_internal_pb2 import (
-    FilePusherStats,
     MetricRecord,
     PollExitResponse,
+    Record,
     RunRecord,
 )
+from wandb.sdk import wandb_print
 from wandb.util import (
     add_import_hook,
-    is_unicode_safe,
     sentry_set_scope,
     to_forward_slash_path,
 )
@@ -78,7 +75,6 @@ from .lib import (
     module,
     proto_util,
     redirect,
-    sparkline,
     telemetry,
 )
 from .lib.exit_hooks import ExitHooks
@@ -100,7 +96,6 @@ if TYPE_CHECKING:
 
 logger = logging.getLogger("wandb")
 EXIT_TIMEOUT = 60
-RUN_NAME_COLOR = "#cdcd00"
 RE_LABEL = re.compile(r"[a-zA-Z0-9_-]+$")
 
 
@@ -269,9 +264,10 @@ class Run(object):
     ]
     _wl: Optional[_WandbSetup]
 
-    _upgraded_version_message: Optional[str]
-    _deleted_version_message: Optional[str]
-    _yanked_version_message: Optional[str]
+    # _upgraded_version_message: Optional[str]
+    # _deleted_version_message: Optional[str]
+    # _yanked_version_message: Optional[str]
+    _check_version: Optional[Record]
 
     _out_redir: Optional[redirect.RedirectBase]
     _err_redir: Optional[redirect.RedirectBase]
@@ -284,9 +280,6 @@ class Run(object):
     _exit_code: Optional[int]
 
     _run_status_checker: Optional[RunStatusChecker]
-    _poll_exit_response: Optional[PollExitResponse]
-
-    _sampled_history: Optional[Dict[str, Union[Sequence[int], Sequence[float]]]]
 
     _use_redirect: bool
     _stdout_slave_fd: Optional[int]
@@ -298,7 +291,6 @@ class Run(object):
     _iface_port: Optional[int]
 
     _attach_id: Optional[str]
-    _final_summary: Optional[Dict[str, Any]]
 
     def __init__(
         self,
@@ -350,17 +342,17 @@ class Run(object):
         self._stderr_slave_fd = None
         self._exit_code = None
         self._exit_result = None
-        self._final_summary = None
-        self._sampled_history = None
         self._jupyter_progress = None
         self._quiet = self._settings._quiet
         if self._settings._jupyter and ipython.in_jupyter():
             self._jupyter_progress = ipython.jupyter_progress_bar()
 
         self._output_writer = None
-        self._upgraded_version_message = None
-        self._deleted_version_message = None
-        self._yanked_version_message = None
+        # self._upgraded_version_message = None
+        # self._deleted_version_message = None
+        # self._yanked_version_message = None
+        self._check_version = None
+        self._printer = wandb_print.Printer()
         self._used_artifact_slots: List[str] = []
 
         # Pull info from settings
@@ -441,7 +433,6 @@ class Run(object):
 
         self._atexit_cleanup_called = False
         self._use_redirect = True
-        self._progress_step = 0
 
         # pid is set so we know if this run object was initialized by this process
         self._init_pid = os.getpid()
@@ -1515,14 +1506,14 @@ class Run(object):
         )
         return visualization
 
-    def _set_upgraded_version_message(self, msg: str) -> None:
-        self._upgraded_version_message = msg
+    # def _set_upgraded_version_message(self, msg: str) -> None:
+    #     self._upgraded_version_message = msg
 
-    def _set_deleted_version_message(self, msg: str) -> None:
-        self._deleted_version_message = msg
+    # def _set_deleted_version_message(self, msg: str) -> None:
+    #     self._deleted_version_message = msg
 
-    def _set_yanked_version_message(self, msg: str) -> None:
-        self._yanked_version_message = msg
+    # def _set_yanked_version_message(self, msg: str) -> None:
+    #     self._yanked_version_message = msg
 
     def _add_panel(
         self, visualize_key: str, panel_type: str, panel_config: dict
@@ -1598,97 +1589,6 @@ class Run(object):
         if not r:
             return ""
         return r.display_name
-
-    def _display_run(self) -> None:
-        project_url = self._get_project_url()
-        run_url = self._get_run_url()
-        sweep_url = self._get_sweep_url()
-        version_str = f"Tracking run with wandb version {wandb.__version__}"
-        if self.resumed:
-            run_state_str = "Resuming run"
-        else:
-            run_state_str = "Syncing run"
-        run_name = self._get_run_name()
-
-        sync_dir = self._settings.sync_dir
-        if self._settings._jupyter:
-            sync_dir = "<code>{}</code>".format(sync_dir)
-        dir_str = "Run data is saved locally in {}".format(sync_dir)
-        if self._settings._jupyter and ipython.in_jupyter():
-            if not wandb.jupyter.maybe_display():
-                # TODO: make settings the source of truth
-                self._quiet = wandb.jupyter.quiet()
-                sweep_line = (
-                    'Sweep page: <a href="{}" target="_blank">{}</a><br/>\n'.format(
-                        sweep_url, sweep_url
-                    )
-                    if sweep_url and not self._quiet
-                    else ""
-                )
-                docs_html = (
-                    ""
-                    if self._quiet
-                    else '(<a href="https://docs.wandb.com/integrations/jupyter.html" target="_blank">docs</a>)'
-                )  # noqa: E501
-                project_html = (
-                    ""
-                    if self._quiet
-                    else f'<a href="{project_url}" target="_blank">Weights & Biases</a>'
-                )
-                ipython.display_html(
-                    """
-                    {} <strong><a href="{}" target="_blank">{}</a></strong> to {} {}.<br/>\n{}
-                """.format(  # noqa: E501
-                        run_state_str,
-                        run_url,
-                        run_name,
-                        project_html,
-                        docs_html,
-                        sweep_line,
-                    )
-                )
-        else:
-            if not self._quiet:
-                wandb.termlog(version_str)
-            wandb.termlog(
-                "{} {}".format(run_state_str, click.style(run_name, fg="yellow"))
-            )
-            emojis = dict(star="", broom="", rocket="")
-            if platform.system() != "Windows" and is_unicode_safe(sys.stdout):
-                emojis = dict(star="⭐️", broom="🧹", rocket="🚀")
-
-            if not self._quiet:
-                wandb.termlog(
-                    "{} View project at {}".format(
-                        emojis.get("star", ""),
-                        click.style(project_url, underline=True, fg="blue"),
-                    )
-                )
-                if sweep_url:
-                    wandb.termlog(
-                        "{} View sweep at {}".format(
-                            emojis.get("broom", ""),
-                            click.style(sweep_url, underline=True, fg="blue"),
-                        )
-                    )
-            wandb.termlog(
-                "{} View run at {}".format(
-                    emojis.get("rocket", ""),
-                    click.style(run_url, underline=True, fg="blue"),
-                )
-            )
-            if not self._quiet:
-                wandb.termlog(dir_str)
-                if not self._settings._offline:
-                    wandb.termlog("Run `wandb offline` to turn off syncing.")
-
-            api = internal.Api()
-            if api.settings().get("anonymous") == "true":
-                wandb.termwarn(
-                    "Do NOT share these links with anyone. They can be used to claim your runs."
-                )
-
-            print("")
 
     def _redirect(
         self,
@@ -1848,94 +1748,34 @@ class Run(object):
             self._output_writer = None
 
     def _on_init(self) -> None:
-        self._show_version_info()
+        self._printer._display_on_init(self._backend.interface)
 
     def _on_start(self) -> None:
-        # TODO: make offline mode in jupyter use HTML
-        if self._settings._offline and not self._quiet:
-            message = (
-                "W&B syncing is set to `offline` in this directory.  ",
-                "Run `wandb online` or set WANDB_MODE=online to enable cloud syncing.",
-            )
-            if self._settings._jupyter and ipython.in_jupyter():
-                ipython.display_html("<br/>\n".join(message))
-            else:
-                for m in message:
-                    wandb.termlog(m)
+
+        self._printer._display_on_start(
+            self._get_project_url(),
+            self._get_run_url(),
+            self._get_sweep_url(),
+            self._get_run_name(),
+            self._settings.sync_dir,
+            self._settings._offline,
+            self._settings._quiet,
+            self._settings._jupyter,
+            self._quiet,
+            self._run_obj,
+            self.resumed,
+        )
 
         if self._settings.save_code and self._settings.code_dir is not None:
             self.log_code(self._settings.code_dir)
-        if self._run_obj and not self._settings._silent:
-            self._display_run()
 
         # TODO(wandb-service) RunStatusChecker not supported yet (WB-7352)
         if self._backend and self._backend.interface and not self._settings._offline:
             self._run_status_checker = RunStatusChecker(self._backend.interface)
         self._console_start()
 
-    def _pusher_print_status(
-        self,
-        progress: FilePusherStats,
-        prefix: bool = True,
-        done: Optional[bool] = False,
-    ) -> None:
-        if self._settings._offline:
-            return
-
-        line = " %.2fMB of %.2fMB uploaded (%.2fMB deduped)\r" % (
-            progress.uploaded_bytes / 1048576.0,
-            progress.total_bytes / 1048576.0,
-            progress.deduped_bytes / 1048576.0,
-        )
-
-        if self._jupyter_progress:
-            percent_done: float
-            if progress.total_bytes == 0:
-                percent_done = 1
-            else:
-                percent_done = progress.uploaded_bytes / progress.total_bytes
-            self._jupyter_progress.update(percent_done, line)
-            if done:
-                self._jupyter_progress.close()
-        elif not self._settings._jupyter:
-            spinner_states = ["-", "\\", "|", "/"]
-
-            line = spinner_states[self._progress_step % 4] + line
-            self._progress_step += 1
-            wandb.termlog(line, newline=False, prefix=prefix)
-
-            if done:
-                dedupe_fraction = (
-                    progress.deduped_bytes / float(progress.total_bytes)
-                    if progress.total_bytes > 0
-                    else 0
-                )
-                if dedupe_fraction > 0.01:
-                    wandb.termlog(
-                        "W&B sync reduced upload amount by %.1f%%             "
-                        % (dedupe_fraction * 100),
-                        prefix=prefix,
-                    )
-                # clear progress line.
-                wandb.termlog(" " * 79, prefix=prefix)
-
-    def _on_finish_progress(self, progress: FilePusherStats, done: bool = None) -> None:
-        self._pusher_print_status(progress, done=done)
-
-    def _wait_for_finish(self) -> PollExitResponse:
-        while True:
-            if self._backend and self._backend.interface:
-                poll_exit_resp = self._backend.interface.communicate_poll_exit()
-            logger.info("got exit ret: %s", poll_exit_resp)
-
-            if poll_exit_resp:
-                done = poll_exit_resp.done
-                pusher_stats = poll_exit_resp.pusher_stats
-                if pusher_stats:
-                    self._on_finish_progress(pusher_stats, done)
-                if done:
-                    return poll_exit_resp
-            time.sleep(0.1)
+    # def _on_finish_progress(self, progress: FilePusherStats, done: bool = None) -> None:
+    # self._pusher_print_status(progress, done=done)
 
     def _on_finish(self) -> None:
         trigger.call("on_finished")
@@ -1951,56 +1791,25 @@ class Run(object):
         self.history._flush()
 
         self._console_stop()  # TODO: there's a race here with jupyter console logging
-        if not self._settings._silent:
-            as_html = self._settings._jupyter and ipython.in_jupyter()
-            if self._backend:
-                pid = self._backend._internal_pid
-                status_str = "Waiting for W&B process to finish, PID {}... ".format(pid)
-            if not self._exit_code:
-                status = "(success)."
-                if as_html:
-                    status = f'<strong style="color:green">{status}</strong>'
-                status_str += status
-            else:
-                status = "(failed {}).".format(self._exit_code)
-                if as_html:
-                    status = f'<strong style="color:red">{status}</strong>'
-                status_str += status
-                if not self._settings._offline:
-                    status_str += " Press ctrl-c to abort syncing."
-            if as_html:
-                sep = "<br/>" if not self._quiet else ""
-                ipython.display_html(sep + status_str)
-            else:
-                print("")
-                wandb.termlog(status_str)
 
         # telemetry could have changed, publish final data
         if self._backend and self._backend.interface:
             self._backend.interface._publish_telemetry(self._telemetry_obj)
 
-        # TODO: we need to handle catastrophic failure better
-        # some tests were timing out on sending exit for reasons not clear to me
-        if self._backend and self._backend.interface:
+            # TODO: we need to handle catastrophic failure better
+            # some tests were timing out on sending exit for reasons not clear to me
             self._backend.interface.publish_exit(self._exit_code)
 
-        # Wait for data to be synced
-        self._poll_exit_response = self._wait_for_finish()
-
         if self._backend and self._backend.interface:
-            ret = self._backend.interface.communicate_get_summary()
-            if ret:
-                self._final_summary = proto_util.dict_from_proto_list(ret.item)
+            self._printer._display_on_finish(
+                self._settings._silent,
+                self._settings._jupyter,
+                self._exit_code,
+                self._settings._offline,
+                self._quiet,
+                self._backend.interface,
+            )
 
-        if self._backend and self._backend.interface:
-            sampled = self._backend.interface.communicate_sampled_history()
-            if sampled:
-                d: Dict[str, Union[Sequence[int], Sequence[float]]] = {}
-                for item in sampled.item:
-                    d[item.key] = (
-                        item.values_float if item.values_float else item.values_int
-                    )
-                self._sampled_history = d
         if self._backend:
             self._backend.cleanup()
 
@@ -2008,203 +1817,19 @@ class Run(object):
             self._run_status_checker.join()
 
     def _on_final(self) -> None:
-        as_html = self._settings._jupyter and ipython.in_jupyter()
-        if as_html:
-            lb = "<br/>\n"
-        else:
-            lb = "\n"
-        # check for warnings and errors, show log file locations
-        final_logs = ""
-        if self._reporter and not self._quiet:
-            warning_lines = self._reporter.warning_lines
-            if warning_lines:
-                final_logs += f"Warnings:{lb}"
-                for line in warning_lines:
-                    final_logs += f"{line}{lb}"
-                if len(warning_lines) < self._reporter.warning_count:
-                    final_logs += f"More warnings...{lb}"
 
-            error_lines = self._reporter.error_lines
-            if error_lines:
-                final_logs += f"Errors:{lb}"
-                for line in error_lines:
-                    final_logs += f"{line}{lb}"
-                if len(error_lines) < self._reporter.error_count:
-                    final_logs += f"More errors...{lb}"
-
-        if not self._quiet:
-            final_logs += self._append_details(final_logs, as_html)
-
-        if self._run_obj:
-            run_url = self._get_run_url()
-            run_name = self._get_run_name()
-            if as_html:
-                final_logs += 'Synced <strong style="color:{}">{}</strong>: <a href="{}" target="_blank">{}</a>{}'.format(
-                    RUN_NAME_COLOR, run_name, run_url, run_url, lb
-                )
-            else:
-                final_logs += "Synced {}: {}{}".format(
-                    click.style(run_name, fg="yellow"),
-                    click.style(run_url, fg="blue"),
-                    lb,
-                )
-
-        if self._settings._offline and not self._quiet:
-            final_logs += f"You can sync this run to the cloud by running:{lb}"
-            final_logs += click.style(
-                f"wandb sync {self._settings.sync_dir}{lb}", fg="yellow"
-            )
-
-        if not self._quiet and (self._settings.log_user or self._settings.log_internal):
-            log_dir = self._settings.log_user or self._settings.log_internal or "."
-            log_dir = log_dir.replace(os.getcwd(), ".")
-            if as_html:
-                log_dir = "<code>{}</code>".format(os.path.dirname(log_dir))
-            final_logs += "Find logs at: {}{}".format(log_dir, lb)
-
-        if as_html:
-            ipython.display_html(final_logs)
-        else:
-            wandb.termlog(final_logs)
-
-        if not self._quiet:
-            self._show_version_info(footer=True)
-            self._show_local_warning()
-
-    def _show_version_info(self, footer: bool = None) -> None:
-        package_problem = False
-        if self._deleted_version_message:
-            wandb.termerror(self._deleted_version_message)
-            package_problem = True
-        elif self._yanked_version_message:
-            wandb.termwarn(self._yanked_version_message)
-            package_problem = True
-        # only display upgrade message if packages are bad or in header
-        if not footer or package_problem:
-            if self._upgraded_version_message:
-                wandb.termlog(self._upgraded_version_message)
-
-    def _append_details(self, logs: str, as_html: bool = False) -> str:
-        if as_html:
-            logs += ipython.TABLE_STYLES
-            logs += '<div class="wandb-row"><div class="wandb-col">\n'
-        logs = self._append_history(logs, as_html)
-        if as_html:
-            logs += '</div><div class="wandb-col">\n'
-        logs = self._append_summary(logs, as_html)
-
-        if as_html:
-            logs += "</div></div>\n"
-        return self._append_files(logs, as_html)
-
-    def _append_summary(self, logs: str, as_html: bool = False) -> str:
-        if self._final_summary:
-            logger.info("rendering summary")
-            max_len = 0
-            summary_rows = []
-            for k, v in sorted(iteritems(self._final_summary)):
-                # arrays etc. might be too large. for now we just don't print them
-                if k.startswith("_"):
-                    continue
-                if isinstance(v, string_types):
-                    if len(v) >= 20:
-                        v = v[:20] + "..."
-                    summary_rows.append((k, v))
-                elif isinstance(v, numbers.Number):
-                    if isinstance(v, float):
-                        v = round(v, 5)
-                    summary_rows.append((k, v))
-                else:
-                    continue
-                max_len = max(max_len, len(k))
-            if not summary_rows:
-                return logs
-            if as_html:
-                summary_table = '<table class="wandb">'
-                for row in summary_rows:
-                    summary_table += "<tr><td>{}</td><td>{}</td></tr>".format(*row)
-                summary_table += "</table>\n"
-                logs += "<h3>Run summary:</h3><br/>" + summary_table
-            else:
-                format_str = "  {:>%s} {}" % max_len
-                summary_lines = "\n".join(
-                    [format_str.format(k, v) for k, v in summary_rows]
-                )
-                logs += f"Run summary:\n{summary_lines}\n\n"
-        return logs
-
-    def _append_history(self, logs: str, as_html: bool = False) -> str:
-        if not self._sampled_history:
-            return logs
-
-        # Only print sparklines if the terminal is utf-8
-        if not is_unicode_safe(sys.stdout):
-            return logs
-
-        logger.info("rendering history")
-        max_len = 0
-        history_rows = []
-        for key in sorted(self._sampled_history):
-            if key.startswith("_"):
-                continue
-            vals = wandb.util.downsample(self._sampled_history[key], 40)
-            if any((not isinstance(v, numbers.Number) for v in vals)):
-                continue
-            line = sparkline.sparkify(vals)
-            history_rows.append((key, line))
-            max_len = max(max_len, len(key))
-        if not history_rows:
-            return logs
-        if as_html:
-            history_table = '<table class="wandb">'
-            for row in history_rows:
-                history_table += "<tr><td>{}</td><td>{}</td></tr>".format(*row)
-            history_table += "</table>"
-            logs += "<h3>Run history:</h3><br/>" + history_table + "<br/>"
-        else:
-            logs += "Run history:\n"
-            history_lines = ""
-            format_str = "  {:>%s} {}\n" % max_len
-            for row in history_rows:
-                history_lines += format_str.format(*row)
-            logs += history_lines.rstrip() + "\n\n"
-        return logs
-
-    def _show_local_warning(self) -> None:
-        if not self._poll_exit_response or not self._poll_exit_response.local_info:
-            return
-
-        if self._settings._offline:
-            return
-
-        if self._settings.is_local:
-            local_info = self._poll_exit_response.local_info
-            latest_version, out_of_date = local_info.version, local_info.out_of_date
-            if out_of_date:
-                wandb.termwarn(
-                    f"Upgrade to the {latest_version} version of W&B Local to get the latest features. Learn more: http://wandb.me/local-upgrade"
-                )
-
-    def _append_files(self, logs: str, as_html: bool = False) -> str:
-        if not self._poll_exit_response or not self._poll_exit_response.file_counts:
-            return logs
-        if self._settings._offline:
-            return logs
-
-        logger.info("logging synced files")
-
-        if self._settings._silent:
-            return logs
-
-        file_str = "Synced {} W&B file(s), {} media file(s), {} artifact file(s) and {} other file(s){}".format(  # noqa:E501
-            self._poll_exit_response.file_counts.wandb_count,
-            self._poll_exit_response.file_counts.media_count,
-            self._poll_exit_response.file_counts.artifact_count,
-            self._poll_exit_response.file_counts.other_count,
-            "\n<br/>" if as_html else "\n",
+        self._printer._display_on_final(
+            self._settings._jupyter,
+            self._quiet,
+            self._settings._silent,
+            self._get_run_url(),
+            self._get_run_name(),
+            self._settings._offline,
+            self._settings.log_user,
+            self._settings.log_internal,
+            self._settings.sync_dir,
+            self._settings.is_local,
         )
-        logs += file_str
-        return logs
 
     def _save_job_spec(self) -> None:
         envdict = dict(python="python3.6", requirements=[],)
